pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "pytest-playwright>=0.7.2",
    "pytest-html>=4.1.1",
    "pylint",
    "pylint-pytest"
]
[tool.pylint.messages_control]
disable = ["missing-module-docstring", "missing-function-docstring"]
[tool.pytest.ini_options]
pythonpath = ["."]